openbsd: use libereadline.so
on OpenBSD, there is several versions of readline library.

the base system has an relatively old version (named libreadline.so), and user
could install a recent library (named libereadline.so) using ports system.

the base version is too old and incompatible with cl-readline. prefer to use the
port version one.